+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Display contributors tab on the about dialog containing the names of all authors of the translation files.
 - The dictionary loading mechanism works far more reliably now.
 - Fixed a bug that would, on certain actions, lead to unwanted behaviour when using the menu items for these actions instead of the shortcut.
+- The last opened file will now be added to the recent documents on start up.
 
 ## Under the Hood